+++ changelog.md
@@ -12,3 +12,10 @@ added - change log
 added - a dictionary to hold map-tile information (add description strings)
 added - movement function
 added - continous play
+
+v3.0 
+Name: map + inventory
+added - empty inventory
+added - second dictonary to store objects in
+added - able to keep items in inventory
+updated - functions created for certain tasks